Moved exception testing to unit test.
